Make it possible to not create service accounts for charts (#23)
This was already possible in the azure-loganalytics-exporter chart and
support for this had already been added to the `_helper.yaml` files.

I've bumped the patch version of each chart since the change is mostly
backwards compatible. If the user has explicitly changed the
serviceAccountName to nil then there service account used changes from
default to the fullname of the chart. But in this case Helm would have
been trying to create a service account with the name default, which
would not work.

---------

Co-authored-by: Markus Blaschke <mblaschke82@gmail.com>